Initial commit
Empty rust project that just prints empty string (simplest valid JSON)

A fork of HWI is installed as a submodule. This fork has a unittest
interface which will run the tests against the "debug" target in this
Rust project. In this way, we can start implementing porting HWI to Rust
while testing against the current HWI Python unittests.